Implement model options
This patch adds a `def_model_option` class method to Model::Base. Model options
represent data that is used when constructing new Vagrant objects, but is not
part of the configuration of those objects. Examples include the autorun option
for config.vm.define and the run option for config.vm.provision.