Add support for OpenStack HTTP settings
This commit adds a `http` attribute to the openstack model which can be set to
a has with values for the `open_timeout`, `read_timeout`, and `proxy` settings.

This settings hash supersedes the `http_read_timeout` and `http_open_timeout`
attributes. These attributes have been removed as it appears the OpenStack
config never had any setting that mapped to these names.